fetchGit: fetch shallow, aka depth=1
This was brought up in # 1 8 3 7 and seems sensible.

The issue talks about implementing a fallback.
One situation in which this fails, is if using fetchGit with a revision
but pointing to a ref like master, because the revision cannot be
exported, if the repo was fetched after the ref was updated.